make new function for validate_perp_fill_possible
