+++ programs/drift/src/controller/orders.rs
@@ -26,7 +26,7 @@ use crate::error::ErrorCode;
 use crate::instructions::{OrderParams, PlaceOrderOptions};
 use crate::load_mut;
 use crate::math::amm_jit::calculate_amm_jit_liquidity;
-use crate::math::auction::{calculate_auction_prices, is_amm_available_liquidity_source};
+use crate::math::auction::calculate_auction_prices;
 use crate::math::casting::Cast;
 use crate::math::constants::{
     BASE_PRECISION_U64, FEE_POOL_TO_REVENUE_POOL_THRESHOLD, FIVE_MINUTE, ONE_HOUR, PERP_DECIMALS,
@@ -1031,18 +1031,7 @@ pub fn fill_perp_order(
         return Ok(0);
     }
 
-    let amm_available = is_amm_available_liquidity_source(
-        &user.orders[order_index],
-        state.min_perp_auction_duration,
-        slot,
-    )?;
-    if !amm_available
-        && makers_and_referrer.0.is_empty()
-        && user.orders[order_index].is_limit_order()
-    {
-        msg!("invalid fill. order is limit order, amm is not available and no makers present");
-        return Err(ErrorCode::ImpossibleFill);
-    }
+    validate_perp_fill_possible(state, user, order_index, slot, makers_and_referrer.0.len())?;
 
     let should_expire_order = should_expire_order_before_fill(user, order_index, now)?;
